LPS-151456 Rename
