+++ modules/apps/frontend-data-set/frontend-data-set-sample-web/src/main/java/com/liferay/frontend/data/set/sample/web/internal/filter/CustomizedCheckBoxFDSFilter.java
@@ -35,7 +35,7 @@ import org.osgi.service.component.annotations.Component;
 	property = "frontend.data.set.name=" + FDSSampleFDSNames.CUSTOMIZED,
 	service = FDSFilter.class
 )
-public class CustomizedSampleCheckBoxFDSFilter extends BaseCheckBoxFDSFilter {
+public class CustomizedCheckBoxFDSFilter extends BaseCheckBoxFDSFilter {
 
 	@Override
 	public List<CheckBoxFDSFilterItem> getCheckBoxFDSFilterItems(
