+++ modules/apps/frontend-data-set/frontend-data-set-sample-web/src/main/java/com/liferay/frontend/data/set/sample/web/internal/filter/CustomizedDateRangeFDSFilter.java
@@ -12,11 +12,14 @@
  * details.
  */
 
-package com.liferay.frontend.data.set.sample.web.internal.view;
+package com.liferay.frontend.data.set.sample.web.internal.filter;
 
+import com.liferay.frontend.data.set.filter.BaseDateRangeFDSFilter;
+import com.liferay.frontend.data.set.filter.DateFDSFilterItem;
+import com.liferay.frontend.data.set.filter.FDSFilter;
 import com.liferay.frontend.data.set.sample.web.internal.constants.FDSSampleFDSNames;
-import com.liferay.frontend.data.set.view.FDSView;
-import com.liferay.frontend.data.set.view.timeline.BaseTimelineFDSView;
+
+import java.util.Calendar;
 
 import org.osgi.service.component.annotations.Component;
 
@@ -24,25 +27,33 @@ import org.osgi.service.component.annotations.Component;
  * @author Javier de Arcos
  */
 @Component(
-	enabled = false,
 	property = "frontend.data.set.name=" + FDSSampleFDSNames.CUSTOMIZED,
-	service = FDSView.class
+	service = FDSFilter.class
 )
-public class CustomizedSampleTimelineFDSView extends BaseTimelineFDSView {
+public class CustomizedDateRangeFDSFilter extends BaseDateRangeFDSFilter {
 
 	@Override
-	public String getDate() {
+	public String getId() {
 		return "date";
 	}
 
 	@Override
-	public String getDescription() {
-		return "description";
+	public String getLabel() {
+		return "Date Range";
 	}
 
 	@Override
-	public String getTitle() {
-		return "title";
+	public DateFDSFilterItem getMaxDateFDSFilterItem() {
+		Calendar calendar = Calendar.getInstance();
+
+		return new DateFDSFilterItem(
+			calendar.get(Calendar.DAY_OF_MONTH),
+			calendar.get(Calendar.MONTH) + 1, calendar.get(Calendar.YEAR));
+	}
+
+	@Override
+	public DateFDSFilterItem getMinDateFDSFilterItem() {
+		return new DateFDSFilterItem(0, 0, 0);
 	}
 
 }
