+++ modules/apps/frontend-data-set/frontend-data-set-sample-web/src/main/java/com/liferay/frontend/data/set/sample/web/internal/view/CustomizedTimelineFDSView.java
@@ -12,14 +12,11 @@
  * details.
  */
 
-package com.liferay.frontend.data.set.sample.web.internal.filter;
+package com.liferay.frontend.data.set.sample.web.internal.view;
 
-import com.liferay.frontend.data.set.filter.BaseDateRangeFDSFilter;
-import com.liferay.frontend.data.set.filter.DateFDSFilterItem;
-import com.liferay.frontend.data.set.filter.FDSFilter;
 import com.liferay.frontend.data.set.sample.web.internal.constants.FDSSampleFDSNames;
-
-import java.util.Calendar;
+import com.liferay.frontend.data.set.view.FDSView;
+import com.liferay.frontend.data.set.view.timeline.BaseTimelineFDSView;
 
 import org.osgi.service.component.annotations.Component;
 
@@ -27,33 +24,25 @@ import org.osgi.service.component.annotations.Component;
  * @author Javier de Arcos
  */
 @Component(
+	enabled = false,
 	property = "frontend.data.set.name=" + FDSSampleFDSNames.CUSTOMIZED,
-	service = FDSFilter.class
+	service = FDSView.class
 )
-public class CustomizedSampleDateRangeFDSFilter extends BaseDateRangeFDSFilter {
+public class CustomizedTimelineFDSView extends BaseTimelineFDSView {
 
 	@Override
-	public String getId() {
+	public String getDate() {
 		return "date";
 	}
 
 	@Override
-	public String getLabel() {
-		return "Date Range";
+	public String getDescription() {
+		return "description";
 	}
 
 	@Override
-	public DateFDSFilterItem getMaxDateFDSFilterItem() {
-		Calendar calendar = Calendar.getInstance();
-
-		return new DateFDSFilterItem(
-			calendar.get(Calendar.DAY_OF_MONTH),
-			calendar.get(Calendar.MONTH) + 1, calendar.get(Calendar.YEAR));
-	}
-
-	@Override
-	public DateFDSFilterItem getMinDateFDSFilterItem() {
-		return new DateFDSFilterItem(0, 0, 0);
+	public String getTitle() {
+		return "title";
 	}
 
 }
