+++ modules/apps/frontend-data-set/frontend-data-set-sample-web/src/main/java/com/liferay/frontend/data/set/sample/web/internal/view/MinimumTableFDSView.java
@@ -35,7 +35,7 @@ import org.osgi.service.component.annotations.Reference;
 	property = "frontend.data.set.name=" + FDSSampleFDSNames.MINIMUM,
 	service = FDSView.class
 )
-public class MinimumSampleTableFDSView extends BaseTableFDSView {
+public class MinimumTableFDSView extends BaseTableFDSView {
 
 	@Override
 	public FDSTableSchema getFDSTableSchema(Locale locale) {
